+++ satella/coding/__init__.py
@@ -14,13 +14,13 @@ from .metaclasses import metaclass_maker, wrap_with, dont_wrap, wrap_property, D
     CopyDocsFrom
 from .misc import update_if_not_none, update_key_if_none, update_attr_if_none, queue_iterator, \
     update_key_if_not_none, source_to_function, update_key_if_true, \
-    get_arguments, call_with_arguments, chain_callables, Closeable
+    get_arguments, call_with_arguments, chain_callables, Closeable, contains
 from .overloading import overload, class_or_instancemethod
 from .recast_exceptions import rethrow_as, silence_excs, catch_exception, log_exceptions, \
     raises_exception
 
 __all__ = [
-    'Closeable',
+    'Closeable', 'contains',
     'overload', 'class_or_instancemethod',
     'update_if_not_none', 'DocsFromParent', 'update_key_if_none', 'queue_iterator',
     'update_attr_if_none', 'update_key_if_not_none', 'source_to_function',