+++ tests/test_coding/test_misc.py
@@ -3,12 +3,22 @@ import gc
 import unittest
 
 from satella.coding import update_key_if_not_none, overload, class_or_instancemethod, \
-    update_key_if_true, get_arguments, call_with_arguments, chain_callables, Closeable
-from satella.coding.structures import HashableMixin
+    update_key_if_true, get_arguments, call_with_arguments, chain_callables, Closeable, \
+    contains
+from satella.coding.structures import HashableMixin, ComparableEnum
 from satella.coding.transforms import jsonify, intify
 
 
 class TestCase(unittest.TestCase):
+
+    def test_contains(self):
+        class CEnum(ComparableEnum):
+            A = 1
+            B = 2
+
+        a = [CEnum.A, CEnum.B]
+        self.assertTrue(contains(2, a))
+        self.assertFalse(contains(3, a))
 
     def test_closeable(self):
         a = {'test': False}