+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.30
+
+* added `enum_value`